refactor(handler): 拆分 MCPToolHandler 以遵循单一职责原则
将 2743 行的 MCPToolHandler 拆分为多个专门的服务：

## 新增服务
- ToolValidator (561 行): 工具验证逻辑
- ToolErrorHandler (300 行): 错误处理
- ToolPreCheckService (223 行): 边界条件预检查
- CozeWorkflowConverter (232 行): Workflow 转换
- ToolSchemaGenerator (91 行): Schema 生成

## 改进
- Handler 从 2743 行减少到 1156 行 (减少 58%)
- 每个服务职责清晰，易于测试和维护
- 保持 API 向后兼容

Fixes #3097
