+++ apps/backend/services/tool/ToolErrorHandler.ts
@@ -1,0 +1,300 @@
+/**
+ * 工具错误处理服务
+ * 提供统一的错误处理和响应格式化功能
+ */
+
+import type { Logger } from "@/Logger.js";
+import { ToolValidator } from "./ToolValidator.js";
+
+/**
+ * 错误响应结构
+ */
+export interface ErrorResponse {
+  code: string;
+  message: string;
+  status: number;
+}
+
+/**
+ * 工具错误处理服务
+ */
+export class ToolErrorHandler {
+  private validator: ToolValidator;
+
+  constructor(private logger: Logger) {
+    this.validator = new ToolValidator();
+  }
+
+  /**
+   * 处理添加工具时的错误
+   */
+  handleAddToolError(error: unknown): ErrorResponse {
+    const errorMessage =
+      error instanceof Error ? error.message : "添加自定义工具失败";
+
+    // 工具类型错误 (400)
+    if (
+      errorMessage.includes("工具类型") ||
+      errorMessage.includes("TOOL_TYPE")
+    ) {
+      return {
+        code: "INVALID_TOOL_TYPE",
+        message: errorMessage,
+        status: 400,
+      };
+    }
+
+    // 缺少必需字段错误 (400)
+    if (
+      errorMessage.includes("必需字段") ||
+      errorMessage.includes("MISSING_REQUIRED_FIELD")
+    ) {
+      return {
+        code: "MISSING_REQUIRED_FIELD",
+        message: errorMessage,
+        status: 400,
+      };
+    }
+
+    // 工具或服务不存在错误 (404)
+    if (
+      errorMessage.includes("不存在") ||
+      errorMessage.includes("NOT_FOUND") ||
+      errorMessage.includes("未找到")
+    ) {
+      return {
+        code: "SERVICE_OR_TOOL_NOT_FOUND",
+        message: errorMessage,
+        status: 404,
+      };
+    }
+
+    // 服务未初始化错误 (503)
+    if (
+      errorMessage.includes("未初始化") ||
+      errorMessage.includes("SERVICE_NOT_INITIALIZED")
+    ) {
+      return {
+        code: "SERVICE_NOT_INITIALIZED",
+        message: errorMessage,
+        status: 503,
+      };
+    }
+
+    // 工具名称冲突错误 (409)
+    if (
+      errorMessage.includes("已存在") ||
+      errorMessage.includes("冲突") ||
+      errorMessage.includes("TOOL_NAME_CONFLICT")
+    ) {
+      return {
+        code: "TOOL_NAME_CONFLICT",
+        message: `${errorMessage}。建议：1) 使用自定义名称；2) 删除现有同名工具后重试`,
+        status: 409,
+      };
+    }
+
+    // 数据验证错误 (400)
+    if (this.validator.isValidationError(errorMessage)) {
+      return {
+        code: "VALIDATION_ERROR",
+        message: this.validator.formatValidationError(errorMessage),
+        status: 400,
+      };
+    }
+
+    // 配置错误 (422)
+    if (
+      errorMessage.includes("配置") ||
+      errorMessage.includes("token") ||
+      errorMessage.includes("API") ||
+      errorMessage.includes("CONFIGURATION_ERROR")
+    ) {
+      return {
+        code: "CONFIGURATION_ERROR",
+        message: `${errorMessage}。请检查：1) 相关配置是否正确；2) 网络连接是否正常；3) 配置文件权限是否正确`,
+        status: 422,
+      };
+    }
+
+    // 资源限制错误 (429)
+    if (
+      errorMessage.includes("资源限制") ||
+      errorMessage.includes("RESOURCE_LIMIT_EXCEEDED")
+    ) {
+      return {
+        code: "RESOURCE_LIMIT_EXCEEDED",
+        message: errorMessage,
+        status: 429,
+      };
+    }
+
+    // 未实现功能错误 (501)
+    if (
+      errorMessage.includes("未实现") ||
+      errorMessage.includes("NOT_IMPLEMENTED")
+    ) {
+      return {
+        code: "TOOL_TYPE_NOT_IMPLEMENTED",
+        message: errorMessage,
+        status: 501,
+      };
+    }
+
+    // 系统错误 (500)
+    return {
+      code: "ADD_CUSTOM_TOOL_ERROR",
+      message: `添加工具失败：${errorMessage}。请稍后重试，如问题持续存在请联系管理员`,
+      status: 500,
+    };
+  }
+
+  /**
+   * 处理更新工具时的错误
+   */
+  handleUpdateToolError(error: unknown): ErrorResponse {
+    const errorMessage =
+      error instanceof Error ? error.message : "更新自定义工具配置失败";
+
+    // 工具不存在错误 (404)
+    if (errorMessage.includes("不存在") || errorMessage.includes("未找到")) {
+      return {
+        code: "TOOL_NOT_FOUND",
+        message: `${errorMessage}。请检查工具名称是否正确`,
+        status: 404,
+      };
+    }
+
+    // 工具类型错误 (400)
+    if (
+      errorMessage.includes("工具类型") ||
+      errorMessage.includes("INVALID_TOOL_TYPE")
+    ) {
+      return {
+        code: "INVALID_TOOL_TYPE",
+        message: errorMessage,
+        status: 400,
+      };
+    }
+
+    // 参数错误 (400)
+    if (errorMessage.includes("不能为空") || errorMessage.includes("无效")) {
+      return {
+        code: "INVALID_REQUEST",
+        message: `${errorMessage}。请提供有效的工具配置数据`,
+        status: 400,
+      };
+    }
+
+    // 配置错误 (422)
+    if (errorMessage.includes("配置") || errorMessage.includes("权限")) {
+      return {
+        code: "CONFIGURATION_ERROR",
+        message: `${errorMessage}。请检查配置文件权限和格式是否正确`,
+        status: 422,
+      };
+    }
+
+    // 未实现功能错误 (501)
+    if (
+      errorMessage.includes("未实现") ||
+      errorMessage.includes("NOT_IMPLEMENTED")
+    ) {
+      return {
+        code: "TOOL_TYPE_NOT_IMPLEMENTED",
+        message: errorMessage,
+        status: 501,
+      };
+    }
+
+    // 系统错误 (500)
+    return {
+      code: "UPDATE_CUSTOM_TOOL_ERROR",
+      message: `更新工具配置失败：${errorMessage}。请稍后重试，如问题持续存在请联系管理员`,
+      status: 500,
+    };
+  }
+
+  /**
+   * 处理删除工具时的错误
+   */
+  handleRemoveToolError(error: unknown): ErrorResponse {
+    const errorMessage =
+      error instanceof Error ? error.message : "删除自定义工具失败";
+
+    // 工具不存在错误 (404)
+    if (errorMessage.includes("不存在") || errorMessage.includes("未找到")) {
+      return {
+        code: "TOOL_NOT_FOUND",
+        message: `${errorMessage}。请检查工具名称是否正确，或刷新页面查看最新的工具列表`,
+        status: 404,
+      };
+    }
+
+    // 参数错误 (400)
+    if (errorMessage.includes("不能为空") || errorMessage.includes("无效")) {
+      return {
+        code: "INVALID_REQUEST",
+        message: `${errorMessage}。请提供有效的工具名称`,
+        status: 400,
+      };
+    }
+
+    // 配置错误 (422)
+    if (errorMessage.includes("配置") || errorMessage.includes("权限")) {
+      return {
+        code: "CONFIGURATION_ERROR",
+        message: `${errorMessage}。请检查配置文件权限和格式是否正确`,
+        status: 422,
+      };
+    }
+
+    // 系统错误 (500)
+    return {
+      code: "REMOVE_CUSTOM_TOOL_ERROR",
+      message: `删除工具失败：${errorMessage}。请稍后重试，如问题持续存在请联系管理员`,
+      status: 500,
+    };
+  }
+
+  /**
+   * 处理工具调用错误
+   */
+  handleToolCallError(error: unknown): ErrorResponse {
+    const errorMessage =
+      error instanceof Error ? error.message : "工具调用失败";
+
+    let errorCode = "TOOL_CALL_ERROR";
+
+    if (errorMessage.includes("不存在")) {
+      errorCode = "SERVICE_OR_TOOL_NOT_FOUND";
+    } else if (
+      errorMessage.includes("未启动") ||
+      errorMessage.includes("未连接")
+    ) {
+      errorCode = "SERVICE_NOT_AVAILABLE";
+    } else if (errorMessage.includes("已被禁用")) {
+      errorCode = "TOOL_DISABLED";
+    } else if (errorMessage.includes("参数验证失败")) {
+      errorCode = "INVALID_ARGUMENTS";
+    } else if (
+      errorMessage.includes("CustomMCP") ||
+      errorMessage.includes("customMCP")
+    ) {
+      errorCode = "CUSTOM_MCP_ERROR";
+    } else if (
+      errorMessage.includes("工作流调用失败") ||
+      errorMessage.includes("API 请求失败")
+    ) {
+      errorCode = "EXTERNAL_API_ERROR";
+    } else if (errorMessage.includes("超时")) {
+      errorCode = "TIMEOUT_ERROR";
+    }
+
+    return {
+      code: errorCode,
+      message: errorMessage,
+      status: 500,
+    };
+  }
+}
