+++ apps/backend/services/tool/ToolPreCheckService.ts
@@ -1,0 +1,223 @@
+/**
+ * 工具预检查服务
+ * 提供工具添加前的边界条件预检查功能
+ */
+
+import type { Logger } from "@/Logger.js";
+import { configManager } from "@xiaozhi-client/config";
+import { ToolValidator } from "./ToolValidator.js";
+
+/**
+ * 预检查结果
+ */
+export interface PreCheckResult {
+  code: string;
+  message: string;
+  status: number;
+}
+
+/**
+ * 工具预检查服务
+ */
+export class ToolPreCheckService {
+  private validator: ToolValidator;
+
+  constructor(private logger: Logger) {
+    this.validator = new ToolValidator();
+  }
+
+  /**
+   * 执行边界条件预检查
+   */
+  performPreChecks(
+    workflow: unknown,
+    customName?: string,
+    customDescription?: string
+  ): PreCheckResult | null {
+    // 检查基础参数
+    const basicCheckResult = this.checkBasicParameters(
+      workflow,
+      customName,
+      customDescription
+    );
+    if (basicCheckResult) return basicCheckResult;
+
+    // 检查系统状态
+    const systemCheckResult = this.checkSystemStatus();
+    if (systemCheckResult) return systemCheckResult;
+
+    // 检查资源限制
+    const resourceCheckResult = this.checkResourceLimits();
+    if (resourceCheckResult) return resourceCheckResult;
+
+    return null;
+  }
+
+  /**
+   * 检查基础参数
+   */
+  private checkBasicParameters(
+    workflow: unknown,
+    customName?: string,
+    customDescription?: string
+  ): PreCheckResult | null {
+    if (!workflow) {
+      return {
+        code: "INVALID_REQUEST",
+        message: "请求体中缺少 workflow 参数",
+        status: 400,
+      };
+    }
+
+    if (typeof workflow !== "object") {
+      return {
+        code: "INVALID_REQUEST",
+        message: "workflow 参数必须是对象类型",
+        status: 400,
+      };
+    }
+
+    if (!Array.isArray(workflow)) {
+      const workflowObj = workflow as Record<string, unknown>;
+
+      if (
+        !workflowObj.workflow_id ||
+        typeof workflowObj.workflow_id !== "string" ||
+        !workflowObj.workflow_id.trim()
+      ) {
+        return {
+          code: "INVALID_REQUEST",
+          message: "workflow_id 不能为空且必须是非空字符串",
+          status: 400,
+        };
+      }
+
+      if (
+        !workflowObj.workflow_name ||
+        typeof workflowObj.workflow_name !== "string" ||
+        !workflowObj.workflow_name.trim()
+      ) {
+        return {
+          code: "INVALID_REQUEST",
+          message: "workflow_name 不能为空且必须是非空字符串",
+          status: 400,
+        };
+      }
+    }
+
+    if (customName !== undefined) {
+      if (typeof customName !== "string") {
+        return {
+          code: "INVALID_REQUEST",
+          message: "customName 必须是字符串类型",
+          status: 400,
+        };
+      }
+
+      if (customName.trim() === "") {
+        return {
+          code: "INVALID_REQUEST",
+          message: "customName 不能为空字符串",
+          status: 400,
+        };
+      }
+
+      if (customName.length > 50) {
+        return {
+          code: "INVALID_REQUEST",
+          message: "customName 长度不能超过50个字符",
+          status: 400,
+        };
+      }
+    }
+
+    if (customDescription !== undefined) {
+      if (typeof customDescription !== "string") {
+        return {
+          code: "INVALID_REQUEST",
+          message: "customDescription 必须是字符串类型",
+          status: 400,
+        };
+      }
+
+      if (customDescription.length > 200) {
+        return {
+          code: "INVALID_REQUEST",
+          message: "customDescription 长度不能超过200个字符",
+          status: 400,
+        };
+      }
+    }
+
+    return null;
+  }
+
+  /**
+   * 检查系统状态
+   */
+  private checkSystemStatus(): PreCheckResult | null {
+    try {
+      const cozeConfig = configManager.getCozePlatformConfig();
+      if (!cozeConfig || !cozeConfig.token) {
+        return {
+          code: "CONFIGURATION_ERROR",
+          message:
+            "未配置扣子API Token。请在系统设置中配置 platforms.coze.token",
+          status: 422,
+        };
+      }
+
+      if (
+        typeof cozeConfig.token !== "string" ||
+        cozeConfig.token.trim() === ""
+      ) {
+        return {
+          code: "CONFIGURATION_ERROR",
+          message: "扣子API Token格式无效。请检查配置中的 platforms.coze.token",
+          status: 422,
+        };
+      }
+    } catch {
+      return {
+        code: "SYSTEM_ERROR",
+        message: "系统配置检查失败，请稍后重试",
+        status: 500,
+      };
+    }
+
+    return null;
+  }
+
+  /**
+   * 检查资源限制
+   */
+  private checkResourceLimits(): PreCheckResult | null {
+    try {
+      const existingTools = configManager.getCustomMCPTools();
+      const maxTools = 100;
+
+      if (existingTools.length >= maxTools) {
+        return {
+          code: "RESOURCE_LIMIT_EXCEEDED",
+          message: `已达到最大工具数量限制 (${maxTools})。请删除一些不需要的工具后重试`,
+          status: 429,
+        };
+      }
+
+      const configSizeEstimate = JSON.stringify(existingTools).length;
+      const maxConfigSize = 1024 * 1024;
+
+      if (configSizeEstimate > maxConfigSize) {
+        return {
+          code: "PAYLOAD_TOO_LARGE",
+          message: "配置文件过大。请删除一些不需要的工具以释放空间",
+          status: 413,
+        };
+      }
+    } catch (error) {
+      this.logger.warn("资源限制检查失败:", error);
+    }
+
+    return null;
+  }
+}
