+++ apps/backend/services/tool/ToolSchemaGenerator.ts
@@ -1,0 +1,91 @@
+/**
+ * 工具 Schema 生成服务
+ * 提供工具输入参数 Schema 生成功能
+ */
+
+import type { CozeWorkflow, WorkflowParameterConfig } from "@/types/coze.js";
+import type { JSONSchema } from "@/types/toolApi.js";
+
+/**
+ * 工具 Schema 生成服务
+ */
+export class ToolSchemaGenerator {
+  /**
+   * 生成输入参数结构
+   */
+  generateInputSchema(
+    workflow: CozeWorkflow,
+    parameterConfig?: WorkflowParameterConfig
+  ): JSONSchema {
+    // 如果提供了参数配置，使用参数配置生成 schema
+    if (parameterConfig && parameterConfig.parameters.length > 0) {
+      return this.generateInputSchemaFromConfig(parameterConfig);
+    }
+
+    // 否则使用默认的基础参数结构
+    return this.generateDefaultInputSchema();
+  }
+
+  /**
+   * 生成默认输入参数结构
+   */
+  private generateDefaultInputSchema(): JSONSchema {
+    return {
+      type: "object",
+      properties: {
+        input: {
+          type: "string",
+          description: "输入内容",
+        },
+      },
+      required: ["input"],
+      additionalProperties: false,
+    };
+  }
+
+  /**
+   * 根据参数配置生成输入参数结构
+   */
+  private generateInputSchemaFromConfig(
+    parameterConfig: WorkflowParameterConfig
+  ): JSONSchema {
+    const properties: Record<string, unknown> = {};
+    const required: string[] = [];
+
+    for (const param of parameterConfig.parameters) {
+      properties[param.fieldName] = {
+        type: param.type,
+        description: param.description,
+      };
+
+      if (param.required) {
+        required.push(param.fieldName);
+      }
+    }
+
+    return {
+      type: "object",
+      properties,
+      required: required.length > 0 ? required : undefined,
+      additionalProperties: false,
+    };
+  }
+
+  /**
+   * 生成工具描述
+   */
+  generateToolDescription(
+    workflow: CozeWorkflow,
+    customDescription?: string
+  ): string {
+    if (customDescription) {
+      return customDescription;
+    }
+
+    if (workflow.description?.trim()) {
+      return workflow.description.trim();
+    }
+
+    return `扣子工作流工具: ${workflow.workflow_name}`;
+  }
+}
