+++ apps/backend/services/tool/index.ts
@@ -1,0 +1,13 @@
+/**
+ * 工具服务模块
+ * 提供工具验证、Schema生成、错误处理、预检查和转换服务
+ */
+
+export { ToolValidator, VALIDATION_REGEX } from "./ToolValidator.js";
+export { ToolSchemaGenerator } from "./ToolSchemaGenerator.js";
+export { ToolErrorHandler, type ErrorResponse } from "./ToolErrorHandler.js";
+export { CozeWorkflowConverter } from "./CozeWorkflowConverter.js";
+export {
+  ToolPreCheckService,
+  type PreCheckResult,
+} from "./ToolPreCheckService.js";
